+++ apps/admin/app/(routes)/(auth)/new-user/page.tsx
