+++ apps/admin/app/(routes)/(auth)/sign-out/page.tsx
@@ -1,4 +1,4 @@
-import { SignOut } from '@/app/sections'
+import { SignOut } from '@/app/sections/auth'
 
 export default async function Page() {
   return <SignOut />
