+++ apps/admin/app/layouts/ContentLayout.tsx
@@ -1,18 +1,23 @@
 'use client'
 
 import { Breadcrumbs, Navbar } from '@/app/sections'
+import React from 'react'
 
 interface ContentLayoutProps {
   title?: string
+  actions?: React.ReactNode
   children: React.ReactNode
 }
 
-const ContentLayout = ({ title, children }: ContentLayoutProps) => {
+const ContentLayout = ({ title, children, actions }: ContentLayoutProps) => {
   return (
     <div>
       <Navbar title={title} />
       <div className='container flex flex-col px-4 pb-8 pt-8 sm:px-8'>
-        <Breadcrumbs />
+        <div className='flex flex-row items-center justify-between'>
+          <Breadcrumbs />
+          {actions}
+        </div>
         {children}
       </div>
     </div>
