+++ apps/admin/app/sections/index.ts
@@ -1,5 +1,3 @@
 export * from './Breadcrumbs'
 export * from './Navbar'
 export * from './Sidebar'
-export * from './SignIn'
-export * from './SignOut'
