+++ apps/admin/app/sections/posts/index.ts
@@ -1,0 +1,1 @@
+export * from './CreatePostButton'
